# spec/openapi/inventory-service-v1.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://127.0.0.1:8085","description":"Generated server url"}],"paths":{"/internal/v1/inventory/return-sold":{"post":{"tags":["inventory-return-controller"],"operationId":"returned","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InventoryCommand"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseInventoryView"}}}}}}},"/internal/v1/inventory/reserve":{"post":{"tags":["inventory-controller"],"operationId":"reserve","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InventoryCommand"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseInventoryView"}}}}}}},"/internal/v1/inventory/release":{"post":{"tags":["inventory-controller"],"operationId":"release","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InventoryCommand"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseInventoryView"}}}}}}},"/internal/v1/inventory/maintenance/release-expired":{"post":{"tags":["inventory-maintenance-controller"],"operationId":"release_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseMapStringInteger"}}}}}}},"/internal/v1/inventory/maintenance/reconcile":{"post":{"tags":["inventory-maintenance-controller"],"operationId":"reconcile","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListInventoryReconciliationView"}}}}}}},"/internal/v1/inventory/confirm":{"post":{"tags":["inventory-controller"],"operationId":"confirm","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InventoryCommand"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseInventoryView"}}}}}}},"/internal/v1/inventory/adjust":{"post":{"tags":["inventory-controller"],"operationId":"adjust","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InventoryCommand"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseInventoryView"}}}}}}},"/internal/v1/inventory/{sku}":{"get":{"tags":["inventory-controller"],"operationId":"get","parameters":[{"name":"sku","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseInventoryView"}}}}}}},"/internal/v1/inventory/references/{reference}":{"get":{"tags":["inventory-reference-controller"],"operationId":"reference","parameters":[{"name":"reference","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseInventoryReferenceView"}}}}}}},"/api/v1/admin/inventory":{"get":{"tags":["inventory-administration-controller"],"operationId":"list","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":100}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListInventoryView"}}}}}}}},"components":{"schemas":{"InventoryCommand":{"type":"object","properties":{"requestId":{"type":"string","minLength":1},"skuId":{"type":"string","minLength":1},"quantity":{"type":"integer","format":"int64"},"referenceType":{"type":"string","minLength":1},"referenceId":{"type":"string","minLength":1}},"required":["referenceId","referenceType","requestId","skuId"]},"ApiResponseInventoryView":{"type":"object","properties":{"code":{"type":"string"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/InventoryView"},"traceId":{"type":"string"},"timestamp":{"type":"string","format":"date-time"}}},"InventoryView":{"type":"object","properties":{"skuId":{"type":"string"},"available":{"type":"integer","format":"int64"},"locked":{"type":"integer","format":"int64"},"sold":{"type":"integer","format":"int64"},"version":{"type":"integer","format":"int64"}}},"ApiResponseMapStringInteger":{"type":"object","properties":{"code":{"type":"string"},"message":{"type":"string"},"data":{"type":"object","additionalProperties":{"type":"integer","format":"int32"}},"traceId":{"type":"string"},"timestamp":{"type":"string","format":"date-time"}}},"ApiResponseListInventoryReconciliationView":{"type":"object","properties":{"code":{"type":"string"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/InventoryReconciliationView"}},"traceId":{"type":"string"},"timestamp":{"type":"string","format":"date-time"}}},"InventoryReconciliationView":{"type":"object","properties":{"id":{"type":"string"},"skuId":{"type":"string"},"expectedAvailable":{"type":"integer","format":"int64"},"actualAvailable":{"type":"integer","format":"int64"},"difference":{"type":"integer","format":"int64"},"status":{"type":"string"},"checkedAt":{"type":"string","format":"date-time"}}},"ApiResponseInventoryReferenceView":{"type":"object","properties":{"code":{"type":"string"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/InventoryReferenceView"},"traceId":{"type":"string"},"timestamp":{"type":"string","format":"date-time"}}},"InventoryReferenceView":{"type":"object","properties":{"referenceId":{"type":"string"},"lockedQuantity":{"type":"integer","format":"int64"},"soldQuantity":{"type":"integer","format":"int64"},"releasedQuantity":{"type":"integer","format":"int64"},"returnedQuantity":{"type":"integer","format":"int64"}}},"ApiResponseListInventoryView":{"type":"object","properties":{"code":{"type":"string"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/InventoryView"}},"traceId":{"type":"string"},"timestamp":{"type":"string","format":"date-time"}}},"ApiResponse":{"type":"object","properties":{"code":{"type":"string","description":"稳定业务错误码"},"message":{"type":"string","description":"可安全展示的消息"},"data":{"description":"业务数据；失败时通常为空"},"traceId":{"type":"string","description":"链路追踪标识"},"timestamp":{"type":"string","format":"date-time"}},"required":["code","data","message","timestamp","traceId"]},"FieldValidationError":{"type":"object","properties":{"field":{"type":"string"},"message":{"type":"string"},"rejectedValue":{"description":"始终为空，避免敏感输入泄漏"}},"required":["field","message","rejectedValue"]},"ValidationErrorResponse":{"type":"object","properties":{"code":{"type":"string"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/FieldValidationError"}},"traceId":{"type":"string"},"timestamp":{"type":"string","format":"date-time"}},"required":["code","data","message","timestamp","traceId"]}},"responses":{"ValidationError":{"description":"参数校验失败","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationErrorResponse"}}}},"Unauthorized":{"description":"未登录或登录已失效","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}},"Forbidden":{"description":"无权执行该操作","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}},"NotFound":{"description":"资源不存在","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}},"Conflict":{"description":"业务状态冲突","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}},"RateLimited":{"description":"请求过于频繁","headers":{"Retry-After":{"description":"客户端重试前至少等待的秒数","schema":{"type":"string","pattern":"[1-9][0-9]*"}}},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}},"DependencyUnavailable":{"description":"依赖服务暂不可用","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}},"InternalError":{"description":"系统内部错误","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}},"parameters":{"X-Request-Id":{"name":"X-Request-Id","in":"header","description":"客户端请求标识；缺失或不安全时由服务生成","required":false,"schema":{"type":"string","maxLength":128,"pattern":"[A-Za-z0-9._-]{1,128}"}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}